Update steps for Amperity
Hi there, this PR will update the steps that you do within Amperity to create a query that will return lists of users (along with those user's attributes) that can be sent to Braze. Please do not merge this right away; want to review with a couple of Amperity folks. I also added two links to Amperity documentation that between right now and May 2nd will 404, but those are the correct URLs for the soon-to-be-refreshed docs site, which will be live on May 2nd. Also not sure how to get you updated screen shots, may need to do that asynchronously, and then you can update in a follow-up PR. - james